% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brier_inference.R
\name{Brier_inference}
\alias{Brier_inference}
\title{Inference for Brier score for cureit objects}
\usage{
Brier_inference(object, newdata = NULL, nboot = 100, times, ...)
}
\arguments{
\item{object}{A cureit object.}

\item{newdata}{A \code{base::data.frame()} or \code{tibble::tibble()} containing all
the original predictors used to create x. Defaults to \code{NULL}.}

\item{nboot}{Number of bootstrap resamples generated.}

\item{times}{Numeric vector of times to obtain survival probability estimates at}

\item{...}{Additional arguments passed to other methods.}
}
\value{
a tibble
}
\description{
Inference for Brier score for cureit objects
}
\examples{
c <- cureit(surv_formula = Surv(ttdeath, death) ~ age + grade, 
cure_formula = ~ age + grade,  data = trial)

Brier.Inference(x = c,time=300)
}
\seealso{
Other cureit() functions: 
\code{\link{broom_methods_cureit}},
\code{\link{cureit}()},
\code{\link{nomogram}()},
\code{\link{predict.cureit}()}
}
\concept{cureit() functions}
