% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{broom_methods_smcure}
\alias{broom_methods_smcure}
\alias{tidy.smcure}
\title{Broom methods for smcure objects}
\usage{
\method{tidy}{smcure}(x, exponentiate = FALSE, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An smcure object created by smcure::smcure()}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
coefficient estimates. Defaults to \code{FALSE}.}

\item{conf.int}{A Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{Level of the confidence interval. Default matches that in
\code{smcure(conf.level=)} (typically, 0.95)}

\item{...}{Additional arguments passed to other methods.}
}
\value{
a tibble
}
\description{
Broom methods for smcure objects
}
\examples{
smcure <- smcure::smcure(Surv(ttdeath, death) ~ marker, cureform = ~marker,
 data = trial, model = "ph")

tidy(smcure)
}
\concept{smcure tidiers}
\concept{smcure() functions}
