% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cureit-package.R
\docType{package}
\name{cureit-package}
\alias{cureit}
\alias{cureit-package}
\title{cureit: A package to build and evaluate cure models}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Karissa Whiting \email{karissa.whiting@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
