% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cureit.R
\name{cureit}
\alias{cureit}
\alias{cureit.formula}
\alias{cureit.default}
\title{Cure Model Regression}
\usage{
\method{cureit}{formula}(
  surv_formula,
  cure_formula,
  data,
  conf.level = 0.95,
  nboot = 100,
  eps = 1e-07,
  ...
)

cureit(object, ...)

\method{cureit}{default}(object, ...)
}
\arguments{
\item{surv_formula}{formula with \code{Surv()} on LHS and covariates on RHS.}

\item{cure_formula}{formula with covariates for cure fraction on RHS}

\item{data}{data frame}

\item{conf.level}{confidence level. Default is 0.95.}

\item{nboot}{number of bootstrap samples used for inference. Default number is 100.}

\item{eps}{convergence criterion for the EM algorithm.}

\item{...}{passed to methods}

\item{object}{input object}
}
\value{
\code{cureit} object. The output includes the following:
\itemize{
\item \code{surv_coefs}
\item \code{cure_coefs}
\item \code{surv_formula}
\item \code{cure_formula}
\item \code{data}
\item \code{conf.level}
\item \code{nboot}
\item \code{eps}
\item \code{surv_xlevels}
\item \code{cure_xlevels}
\item \code{tidy}
\item \code{smcure}
\item \code{surv_blueprint}
\item \code{cure_blueprint}
\item \code{blueprint}
}
}
\description{
Cure Model Regression
}
\examples{

cureit_obj <- cureit(surv_formula = Surv(ttdeath, death) ~ age + grade, 
cure_formula = ~ age + grade,  data = trial, nboot = 10)

# pulling survival coeffients
cureit_obj$surv_coefs

# pulling cure coefficients
cureit_obj$coefs

# `tidy` object of survival model output
cureit_obj$tidy$df_surv

# `tidy` object of cure model output
cureit_obj$tidy$df_cure
}
\seealso{
Other cureit() functions: 
\code{\link{Brier_inference_bootstrap}()},
\code{\link{broom_methods_cureit}},
\code{\link{nomogram}()},
\code{\link{predict.cureit}()}
}
\concept{cureit() functions}
