% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_index.R
\name{.calc_k_index}
\alias{.calc_k_index}
\title{K-index estimation}
\usage{
.calc_k_index(object)
}
\arguments{
\item{object}{a cureit model object or a coxph model object}
}
\description{
K-index estimation
}
\examples{

p <- cureit(surv_formula = Surv(ttdeath, death) ~ age, 
   cure_formula = ~ age,
   data = trial) 
.calc_k_index(p)

cox_p <- survival::coxph(Surv(ttdeath, death) ~ age,
   data = trial)
   
.calc_k_index(cox_p)

}
