% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_index.R
\name{k_index}
\alias{k_index}
\title{K-index estimation}
\usage{
k_index(object)
}
\arguments{
\item{object}{a cureit model object}
}
\description{
K-index estimation
}
\examples{
p <- cureit(surv_formula = Surv(ttdeath, death) ~ age, 
   cure_formula = ~ age,
   data = trial) 
k_index(p)
}
