% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Survival after malignant melanoma}
\format{
This data frame contains the following columns:
\describe{
\item{id}{a numeric vector, patient code.}
\item{status}{a numeric vector code, survival status; 1: dead from melanoma, 0: alive or dead from other cause.}
\item{days}{a numeric vector, observation time in days.}
\item{years}{a numeric vector, observation time in years.}
\item{ulceration}{a numeric vector code, ulceration; 1: present, 2: absent.}
\item{tumor_thickness}{a numeric vector, tumor thickness (1/100 mm).}
\item{sex}{a numeric vector code; 1: female, 2: male.}
}
}
\usage{
melanoma
}
\description{
Data frame has 205 rows and 7 columns.
It contains clinical variables and survival outcome data relating to the survival of patients after an operation for
malignant melanoma. It was collected at Odense University Hospital by K.T. Drzewiecki and originally sourced from the {ISwR} Package
}
\keyword{datasets}
