% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram_methods.R
\name{nomogram.cureit}
\alias{nomogram.cureit}
\title{Nomogram}
\usage{
\method{nomogram}{cureit}(x, survival = TRUE, cure = TRUE, time = NULL, angle = 0, ...)
}
\arguments{
\item{x}{a cure model fit}

\item{survival}{Logical indicating whether or not to output the nomogram
based on the survival submodel. Defaults to \code{TRUE}.}

\item{cure}{Logical indicating whether or not to output the nomogram
based on the cured probability submodel. Defaults to \code{TRUE}.}

\item{time}{Numeric vector of times at which to obtain survival probability estimates}

\item{angle}{Angle to shift labels}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Nomogram
}
\concept{cureit}
