% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram_methods.R
\name{nomogram}
\alias{nomogram}
\title{Nomogram methods for cure models}
\usage{
nomogram(x, ...)
}
\arguments{
\item{x}{a cure model fit}

\item{...}{Additional arguments passed to other methods.}

\item{survival}{Logical indicating whether or not to output the nomogram
based on the survival submodel. Defaults to \code{TRUE}.}

\item{cure}{Logical indicating whether or not to output the nomogram
based on the cured probability submodel. Defaults to \code{TRUE}.}

\item{time}{Numeric vector of times at which to obtain survival probability estimates}

\item{angle}{Angle to shift labels}
}
\description{
Nomogram methods for cure models
}
\seealso{
\link{nomogram.cureit}
}
\keyword{internal}
