% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram_methods.R
\name{nomogram}
\alias{nomogram}
\title{Nomogram for cureit objects}
\usage{
nomogram(x, survival = TRUE, cure = TRUE, time = NULL, angle = 0, ...)
}
\arguments{
\item{x}{cureit object}

\item{survival}{Logical indicating whether or not to output the nomogram
based on the survival submodel. Defaults to \code{TRUE}.}

\item{cure}{Logical indicating whether or not to output the nomogram
based on the cured probability submodel. Defaults to \code{TRUE}.}

\item{time}{Numeric vector of times to obtain survival probability estimates at}

\item{angle}{angle of text labels}

\item{...}{Additional arguments passed to other methods.}
}
\value{
a tibble
}
\description{
Nomogram for cureit objects
}
\examples{
c <- cureit(surv_formula = Surv(ttdeath, death) ~ age + grade, 
cure_formula = ~ age + grade,  data = trial)

nomogram(x = c,time=300)
}
\seealso{
Other cureit() functions: 
\code{\link{broom_methods_cureit}},
\code{\link{cureit}()},
\code{\link{predict.cureit}()}
}
\concept{cureit() functions}
