% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_methods.R
\name{tbl_regression.cureit}
\alias{tbl_regression.cureit}
\title{gtsummary methods}
\usage{
\method{tbl_regression}{cureit}(x, tidy_fun = cureit::tidy, ...)
}
\arguments{
\item{x}{(\code{cureit})\cr
\code{cureit::cureit()} regression object}

\item{tidy_fun}{(\code{function})\cr
Tidier function for the model. Default is \code{cureit::tidy()}.}

\item{...}{\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}Additional arguments passed to \code{\link[broom.helpers:tidy_plus_plus]{broom.helpers::tidy_plus_plus()}}.
See \code{?tidy_plus_plus_dots} for details.}

\item{type}{not used}
}
\value{
gtsummary table or data frame of results
}
\description{
This function is an S3 methods for working with \code{\link[=cureit]{cureit()}} model
results.
\itemize{
\item \code{tbl_regression.cureit()}: This function sets the cureit tidier for \code{cureit()} models.
}
}
\examples{
cureit(surv_formula = Surv(days, status) ~ ulceration,
  cure_formula = ~ ulceration, data = melanoma) |>
  gtsummary::tbl_regression() |>
  gtsummary::as_gt()
  
}
