% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brier_inference.R
\name{Brier_inference}
\alias{Brier_inference}
\title{Inference for Brier score for cureit objects}
\usage{
Brier_inference(object, newdata = NULL, nboot = 100, times, ...)
}
\arguments{
\item{object}{A cureit object.}

\item{newdata}{A \code{base::data.frame()} or \code{tibble::tibble()} containing all
the original predictors used to create x. Defaults to \code{NULL}.}

\item{nboot}{Number of bootstrap resamples generated.}

\item{times}{Numeric vector of times to obtain survival probability estimates at}

\item{...}{Additional arguments passed to other methods.}
}
\value{
a tibble
}
\description{
Inference for Brier score for cureit objects
}
\examples{
trial <- na.omit(trial)
p <- cureit(
  surv_formula = Surv(ttdeath, death) ~ age,
  cure_formula = ~age,
  data = trial
)

times <- seq(5, 24, 0.5)
bootbrier <- Brier_inference(p, times = times)
plot(times, bootbrier$brier, type = "s", xlab = "Time", ylab = "Brier score")
lines(times, bootbrier$brier_2.5, type = "s", lty = 2)
lines(times, bootbrier$brier_97.5, type = "s", lty = 2)
lines(times, bootbrier$brier_cox, type = "s", col = "red")
lines(times, bootbrier$brier_cox_2.5, type = "s", lty = 2, col = "red")
lines(times, bootbrier$brier_cox_97.5, type = "s", lty = 2, col = "red")
legend("topleft", c("Cure model", "Cox model"), col = c("black", "red"), lty = 1)

}
\seealso{
Other cureit() functions: 
\code{\link{broom_methods_cureit}},
\code{\link{cureit}()},
\code{\link{nomogram}()},
\code{\link{predict.cureit}()}
}
\concept{cureit() functions}
