% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{broom_methods_cureit}
\alias{broom_methods_cureit}
\alias{tidy.cureit}
\title{Broom methods for cureit objects}
\usage{
\method{tidy}{cureit}(x, exponentiate = FALSE, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{exponentiate}{Logical indicating whether or not to exponentiate the
coefficient estimates. Defaults to \code{FALSE}.}

\item{conf.level}{Level of the confidence interval. Default matches that in
\code{smcure(conf.level=)} (typically, 0.95)}
}
\value{
a tibble
}
\description{
Broom methods for cureit objects
}
\examples{
cureit <- cureit(Surv(ttdeath, death_cr) ~ age + grade, trial)

tidy(cureit)
}
\seealso{
Other cureit() functions: 
\code{\link{cureit}()},

\code{\link{nomogram_methods_cureit}},
\code{\link{predict.cureit}()}

}
\concept{cureit tidiers}
\concept{cureit() functions}
