% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{broom_methods_cureit}
\alias{broom_methods_cureit}
\alias{tidy.cureit}
\title{Broom methods for cureit objects}
\usage{
\method{tidy}{cureit}(x, exponentiate = FALSE, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An smcure object created by smcure::smcure()}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
coefficient estimates. Defaults to \code{FALSE}.}

\item{conf.int}{A Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{Level of the confidence interval. Default matches that in
\code{smcure(conf.level=)} (typically, 0.95)}

\item{...}{Additional arguments passed to other methods.}
}
\value{
a tibble
}
\description{
Broom methods for cureit objects
}
\examples{
cureit <- cureit(surv_formula = Surv(ttdeath, death) ~ marker, cure_formula = ~marker,
 data = trial)

tidy(cureit)
}
\seealso{
Other cureit() functions: 
\code{\link{cureit}()},
\code{\link{nomogram}()},
\code{\link{predict.cureit}()}
}
\concept{cureit tidiers}
\concept{cureit() functions}
