% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{broom_methods_smcure}
\alias{broom_methods_smcure}
\alias{tidy.smcure}
\title{Broom methods for smcure objects}
\usage{
\method{tidy}{smcure}(x, exponentiate = FALSE, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{exponentiate}{Logical indicating whether or not to exponentiate the
coefficient estimates. Defaults to \code{FALSE}.}

\item{conf.level}{Level of the confidence interval. Default matches that in
\code{smcure(conf.level=)} (typically, 0.95)}
}
\value{
a tibble
}
\description{
Broom methods for smcure objects
}
\examples{
smcure <- smcure::smcure(Surv(ttdeath, death_cr) ~ age,
cureform =  ~age,
 data = trial)

tidy(smcure)
}
\concept{smcure tidiers}
\concept{smcure() functions}
