% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cureit-package.R
\docType{package}
\name{cureit-package}
\alias{cureit-package}
\alias{{pkgname}-package}
\title{cureit: A package to build and evaluate cure models}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.karissawhiting.com/cureit/}
}

}
\author{
\strong{Maintainer}: Karissa Whiting \email{karissa.whiting@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Teng Fei (\href{https://orcid.org/0000-0001-7888-1715}{ORCID})
}

}
\keyword{internal}
