% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cureit.R
\name{cureit}
\alias{cureit}
\alias{cureit.formula}
\alias{cureit.default}
\title{Cure Model Regression}
\usage{
\method{cureit}{formula}(
  surv_formula,
  cure_formula,
  data,
  conf.level = 0.95,
  nboot = 100,
  eps = 1e-07,
  ...
)

cureit(object, ...)

\method{cureit}{default}(object, ...)
}
\arguments{
\item{surv_formula}{formula with \code{Surv()} on LHS and covariates on RHS.}

\item{cure_formula}{formula with covariates for cure fraction on RHS}

\item{data}{data frame}

\item{conf.level}{confidence level. Default is 0.95.}

\item{nboot}{number of bootstrap samples used for inference.}

\item{eps}{convergence criterion for the EM algorithm.}

\item{...}{passed to methods}

\item{object}{input object}
}
\value{
cureit object
}
\description{
Cure Model Regression
}
\examples{
cureit(surv_formula = Surv(ttdeath, death) ~ age + grade, 
cure_formula = ~ age + grade,  data = trial)
}
\seealso{
Other cureit() functions: 
\code{\link{broom_methods_cureit}},
\code{\link{nomogram}()},
\code{\link{predict.cureit}()}
}
\concept{cureit() functions}
