% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cureit.R
\name{cureit}
\alias{cureit}
\alias{cureit.formula}
\alias{cureit.default}
\title{Cure Model Regression}
\usage{
\method{cureit}{formula}(surv_formula, cure_formula, data, conf.level = 0.95, nboot = 100, ...)

cureit(x, ...)

\method{cureit}{default}(x, ...)
}
\arguments{
\item{data}{data frame}

\item{conf.level}{confidence level. Default is 0.95.}

\item{nboot}{number of bootstrap samples used for inference.}

\item{...}{passed to methods}

\item{x}{input object}

\item{formula}{formula with \code{Surv()} on LHS and covariates on RHS.}
}
\value{
cureit object
}
\description{
Cure Model Regression
}
\examples{
cureit(Surv(ttdeath, death_cr) ~ age + grade, trial)
}
\seealso{
Other cureit() functions: 
\code{\link{broom_methods_cureit}},
\code{\link{nomogram_methods_cureit}}
}
\concept{cureit() functions}
