% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_index.R
\name{.calc_k_index}
\alias{.calc_k_index}
\title{K-index estimation}
\usage{
.calc_k_index(object, newdata = NULL)
}
\arguments{
\item{object}{a cureit model object or a coxph model object}

\item{newdata}{A \code{base::data.frame()} or \code{tibble::tibble()} containing all
the original predictors used to create object. Defaults to \code{NULL}.}
}
\description{
K-index estimation
}
\examples{

p <- cureit(surv_formula = Surv(ttdeath, death) ~ age, 
   cure_formula = ~ age,
   data = trial) 
.calc_k_index(p)

cox_p <- survival::coxph(Surv(ttdeath, death) ~ age,
   data = trial)

# Need to debug this
# .calc_k_index(cox_p)

}
