% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_index.R
\name{k_index}
\alias{k_index}
\title{K-index estimation}
\usage{
k_index(object, newdata = NULL, ncv = NULL)
}
\arguments{
\item{object}{a cureit model object}

\item{newdata}{A \code{base::data.frame()} or \code{tibble::tibble()} containing all
the original predictors used to create object. Defaults to \code{NULL}.}

\item{ncv}{number of folds for cross validation. Defaults to \code{NULL}.
If specified, \code{newdata} will be ignored and cross-validated k-index will be reported.}
}
\description{
K-index estimation
}
\examples{
p <- cureit(surv_formula = Surv(ttdeath, death) ~ age, 
   cure_formula = ~ age,
   data = trial) 
k_index(p)
}
