% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{predict.cureit}
\alias{predict.cureit}
\title{Predicted survival probability by cure model}
\usage{
\method{predict}{cureit}(
  object,
  times = NULL,
  probs = NULL,
  newdata = NULL,
  method = "prob",
  brier = FALSE,
  cox = FALSE,
  ...
)
}
\arguments{
\item{object}{A cureit object.}

\item{times}{Numeric vector of times to obtain survival probability estimates at}

\item{probs}{Numeric vector of quantiles to obtain estimates at}

\item{newdata}{A \code{base::data.frame()} or \code{tibble::tibble()} containing all
the original predictors used to create x. Defaults to \code{NULL}.}

\item{method}{Output format of predicted values: "lp" (linear predictor) or "prob" (predicted probabilities).}

\item{brier}{Boolean indicator of calculating the Brier scores at specified \code{times}.}

\item{cox}{Boolean indicator of fitting the Cox model for the training data and calculating the Brier scores at specified \code{times} for \code{newdata}.}

\item{...}{passed to methods}
}
\value{
named list of prediction estimates
}
\description{
Predicted survival probability by cure model
}
\examples{
p <- cureit(surv_formula = Surv(ttdeath, death) ~ age, 
   cure_formula = ~ age,
   data = trial) 
   
pred <- predict(p, times = seq(5,24,0.5),
 newdata = trial[complete.cases(trial), ], brier=TRUE,cox=TRUE)

plot(seq(5,24,0.5),pred$brier,type="S",pch=1)
lines(seq(5,24,0.5),pred$brier_cox,type="S",col="red",pch=3)
legend("topright",c("Cure model","Cox model"),
col=c("black","red"),lty=1,pch=c(1,3))

  
}
\seealso{
Other cureit() functions: 
\code{\link{Brier_inference}()},
\code{\link{broom_methods_cureit}},
\code{\link{cureit}()},
\code{\link{nomogram}()}
}
\concept{cureit() functions}
