% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{predict.cureit}
\alias{predict.cureit}
\title{Predicted survival probability by cure model}
\usage{
\method{predict}{cureit}(
  object,
  times = NULL,
  probs = NULL,
  newdata = NULL,
  method = "prob",
  ...
)
}
\arguments{
\item{object}{input object}

\item{times}{Numeric vector of times to obtain survival probability estimates at}

\item{probs}{Numeric vector of quantiles to obtain estimates at}

\item{newdata}{A \code{base::data.frame()} or \code{tibble::tibble()} containing all
the original predictors used to create x. Defaults to \code{NULL}.}

\item{method}{Output format of predicted values: "lp" (linear predictor) or "prob" (predicted probabilities).}

\item{...}{passed to methods}
}
\value{
named list of prediction estimates
}
\description{
Predicted survival probability by cure model
}
\examples{
p <- cureit(surv_formula = Surv(ttdeath, death) ~ age, 
   cure_formula = ~ age,
   data = trial) \%>\%
   predict(times = 12, newdata = trial[1:10, ])
  
}
\seealso{
Other cureit() functions: 
\code{\link{broom_methods_cureit}},
\code{\link{cureit}()},
\code{\link{nomogram_methods_cureit}}
}
\concept{cureit() functions}
